requirements.txt
Flask
gunicorn
numpy
pandas
azure.storage